Revert "IMS-VT: Modification of data usage aidl"
This reverts commit 3d008117add89bb421f9b5e6610c07cec407a3cf.

+++ telecomm/java/android/telecom/Connection.java
@@ -675,7 +675,7 @@ public abstract class Connection implements IConferenceable {
          *
          * @param dataUsage The updated data usage.
          */
-        public void changeCallDataUsage(long dataUsage) {
+        public void changeCallDataUsage(int dataUsage) {
             if (mVideoCallback != null) {
                 try {
                     mVideoCallback.changeCallDataUsage(dataUsage);
